Fix LinkedIn auth flow, close #33
This changes the way that the state parameter is dealt with.
Previously, `GetState` was defined twice: once in `gothic.go`, and again
in the example (`examples/main.go`). This redefinition changed the
behaviour of `GetState` from returning a string to returning the value
of the state parameter from the URL query string. That means that the
state parameter wasn't being set by the client during the auth flow.
That's okay for most providers, since sending a state parameter is
optional... Except for with LinkedIn.

Now, there is a `SetState` and `GetState` function. `SetState` returns a
string to be used when setting the `state` parameter as a client during
the auth flow, and `GetState` returns the value of the `state` parameter
in the query string returned by the provider.