Make no Auth URL error message uniform in all providers
I noticed some of the error messages had a typo, and they were
different for some providers. This is my first attempt to make them
all the same.

Some error messages are listing the providers as well, I am happy to
do that in the future if you think this is a good direction for the
code.