Fix provider deduction from existing sessions so that an argument is not needed on the authentication redirect page.
This has been broken since https://github.com/markbates/goth/commit/78433fec1a8fed110af91e79c94be2d835afe637.

Also moving the deduction code to be a fallback: if a provider is explicitly
specified it should not be overridden by the existing session. I believe
this is why the Logout() problems with multiple providers existed in the
first place. If deduction wins over explicit parameters, then once you
wrongly click one provider, the cookie will force you to use that
authentication mechanism even if you've navigated to a URL that
explicitly states another provider's name.

Fixed the test harness to check session key names (this was a bug in the
test harness) and added a test to verify that this change works.