Refactoring `providers/slack.Provider.FetchUser` to use the initially fetched data as a "basic" profile and only fetch the other, more "expansive", user-data if the proper scope has been included in the provider
This improves the Slack provider to only perform an operation when its
able to (has the scope) and allows for users of the provider to choose
whether they want to use the `users:read` scope, which is a scope that
Slack considers to be advanced enough to "warn" about in Slack apps.

Also, this enables the Slack provider to be used with the newer
"Sign in with Slack" flow and potentially ask for more scopes/permission
"progressively".