openidconnect: allow overriding redirect_uri and passing code_verifier
Similar to #459 this PR extends openidconnect provider's
`Session.Authorize` to allow passage of `redirect_uri` and
`code_verifier` params.

Gitea will need these in order to enable support for PKCE in
openidconnect.

It is likely that other providers will also need the addition of PKCE
support however, the design of this may need some more thought.

Signed-off-by: Andrew Thornton <art27@cantab.net>